+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2024-04-18
+## [0.1.4] - 2025-05-19
+
+### Changed
+
+- Updated cipher constructors to use `RandomBytes` function instead of `CryptoModule`
+  - Changed from `new NodeCbcCipher(cryptoModule)` to `new NodeCbcCipher(randomBytes)`
+  - Changed from `new NodeGcmCipher(cryptoModule)` to `new NodeGcmCipher(randomBytes)`
+
+## [0.1.2] - 2025-05-18
 
 ### Fixed